Move file management to a separate page. #21
This avoids bugs when suddenly swapping state on certain pages and will
help with a more clear UX.
