+++ src/routes/cross_section/+layout.svelte
@@ -1,9 +1,14 @@
 <script lang="ts">
-  import { FileManager } from "$lib/files";
-  import { files, currentFile, state } from "./data";
+  import { currentFile } from "./data";
+  import { base } from "$app/paths";
 </script>
 
-<FileManager {files} {currentFile} {state} xlsxImporter={null} />
+<div>
+  <a href="{base}/cross_section/files">Manage files</a>
+  <span>
+    Editing file <u>{$currentFile}</u>
+  </span>
+</div>
 
 <hr />
