+++ src/routes/cross_section/files/+layout@.svelte
@@ -1,0 +1,3 @@
+<!-- Avoid cross_section/layout.svelte, because we don't need a link to this page -->
+<hr />
+<slot />
