+++ src/routes/cross_section/files/+page.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+  import { FileManager } from "$lib/files";
+  import { files, currentFile, state } from "../data";
+</script>
+
+<FileManager {files} {currentFile} {state} />
