+++ src/routes/route_check/files/+layout@.svelte
@@ -1,0 +1,3 @@
+<!-- Avoid route_check/layout.svelte, because we don't need a link to this page -->
+<hr />
+<slot />
